+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@types/styled-components": "^5.1.26",
     "bootstrap": "^5.2.3",
+    "date-and-time": "^2.4.2",
     "immer": "^9.0.17",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
